+++ frontend/src/app/main/ui/viewer/shapes.cljs
@@ -61,14 +61,16 @@
     (let [dest-frame-id              (:destination interaction)
           dest-frame                 (get objects dest-frame-id)
           relative-to-id             (if (= :manual (:overlay-pos-type interaction))
-                                       (:id shape) ;; manual interactions are allways from "self"
+                                       (if (= (:type shape) :frame) ;; manual interactions are always from "self"
+                                         (:frame-id shape)
+                                         (:id shape))
                                        (:position-relative-to interaction))
           relative-to-shape          (or (get objects relative-to-id) base-frame)
           close-click-outside        (:close-click-outside interaction)
           background-overlay         (:background-overlay interaction)
           overlays-ids               (set (map :id overlays))
           relative-to-base-frame     (find-relative-to-base-frame relative-to-shape objects overlays-ids base-frame)
-          position                   (ctsi/calc-overlay-position interaction
+          [position snap-to]         (ctsi/calc-overlay-position interaction
                                                                  shape
                                                                  objects
                                                                  relative-to-shape
@@ -78,6 +80,7 @@
       (when dest-frame-id
         (st/emit! (dv/open-overlay dest-frame-id
                                    position
+                                   snap-to
                                    close-click-outside
                                    background-overlay
                                    (:animation interaction)))))
@@ -86,12 +89,14 @@
     (let [dest-frame-id              (:destination interaction)
           dest-frame                 (get objects dest-frame-id)
           relative-to-id             (if (= :manual (:overlay-pos-type interaction))
-                                       (:id shape) ;; manual interactions are allways from "self"
-                                       (:position-relative-to interaction))
+                                       (if (= (:type shape) :frame) ;; manual interactions are always from "self"
+                                         (:frame-id shape)
+                                         (:id shape))
+                                        (:position-relative-to interaction))
           relative-to-shape          (or (get objects relative-to-id) base-frame)
           overlays-ids               (set (map :id overlays))
           relative-to-base-frame     (find-relative-to-base-frame relative-to-shape objects overlays-ids base-frame)
-          position                   (ctsi/calc-overlay-position interaction
+          [position snap-to]         (ctsi/calc-overlay-position interaction
                                                                  shape
                                                                  objects
                                                                  relative-to-shape
@@ -104,6 +109,7 @@
       (when dest-frame-id
         (st/emit! (dv/toggle-overlay dest-frame-id
                                      position
+                                     snap-to
                                      close-click-outside
                                      background-overlay
                                      (:animation interaction)))))
@@ -136,29 +142,49 @@
       (st/emit! (dv/close-overlay frame-id)))
 
     :toggle-overlay
-    (let [frame-id            (:destination interaction)
-          position            (:overlay-position interaction)
-          close-click-outside (:close-click-outside interaction)
-          background-overlay  (:background-overlay interaction)]
-      (when frame-id
-        (st/emit! (dv/toggle-overlay frame-id
+    (let [dest-frame-id              (:destination interaction)
+          dest-frame                 (get objects dest-frame-id)
+          relative-to-id             (if (= :manual (:overlay-pos-type interaction))
+                                       (if (= (:type shape) :frame) ;; manual interactions are always from "self"
+                                         (:frame-id shape)
+                                         (:id shape))
+                                        (:position-relative-to interaction))
+          relative-to-shape          (or (get objects relative-to-id) base-frame)
+          overlays-ids               (set (map :id overlays))
+          relative-to-base-frame     (find-relative-to-base-frame relative-to-shape objects overlays-ids base-frame)
+          [position snap-to]         (ctsi/calc-overlay-position interaction
+                                                                 shape
+                                                                 objects
+                                                                 relative-to-shape
+                                                                 relative-to-base-frame
+                                                                 dest-frame
+                                                                 frame-offset)
+
+          close-click-outside        (:close-click-outside interaction)
+          background-overlay         (:background-overlay interaction)]
+      (when dest-frame-id
+        (st/emit! (dv/toggle-overlay dest-frame-id
                                      position
+                                     snap-to
                                      close-click-outside
                                      background-overlay
                                      (:animation interaction)))))
+
 
     :close-overlay
     (let [dest-frame-id              (:destination interaction)
           dest-frame                 (get objects dest-frame-id)
           relative-to-id             (if (= :manual (:overlay-pos-type interaction))
-                                       (:id shape) ;; manual interactions are allways from "self"
+                                       (if (= (:type shape) :frame) ;; manual interactions are always from "self"
+                                         (:frame-id shape)
+                                         (:id shape))
                                        (:position-relative-to interaction))
           relative-to-shape          (or (get objects relative-to-id) base-frame)
           close-click-outside        (:close-click-outside interaction)
           background-overlay         (:background-overlay interaction)
           overlays-ids               (set (map :id overlays))
           relative-to-base-frame     (find-relative-to-base-frame relative-to-shape objects overlays-ids base-frame)
-          position                   (ctsi/calc-overlay-position interaction
+          [position snap-to]         (ctsi/calc-overlay-position interaction
                                                                  shape
                                                                  objects
                                                                  relative-to-shape
@@ -168,6 +194,7 @@
       (when dest-frame-id
         (st/emit! (dv/open-overlay dest-frame-id
                                    position
+                                   snap-to
                                    close-click-outside
                                    background-overlay
                                    (:animation interaction)))))
@@ -266,6 +293,9 @@
                 svg-element?       (and (= :svg-raw (:type shape))
                                         (not= :svg (get-in shape [:content :tag])))
 
+                ;; _ (js/console.log "======" (:name shape))
+                ;; _ (js/console.log "shape" (clj->js shape))
+                ;; _ (js/console.log "frame-offset" (clj->js frame-offset))
                 on-pointer-down
                 (mf/use-fn (mf/deps shape base-frame frame-offset objects)
                            #(on-pointer-down % shape base-frame frame-offset objects overlays))